Clean up implementation and tests

+++ test/histories.js
@@ -1,8 +1,10 @@
 import createHashHistory from 'history/lib/createHashHistory';
 
-export function createHashHistoryWithoutKey() {
-  // Avoid persistence of stored data from previous tests.
-  window.sessionStorage.clear();
+export function createHashHistoryWithoutKey({ resetState = true } = {}) {
+  if (resetState) {
+    // Avoid persistence of stored data from previous tests.
+    window.sessionStorage.clear();
+  }
 
   return createHashHistory({ queryKey: false });
 }
